fix: optimize code and remove the backend api calls
