+++ swapstreet/app/auth/sign-in/page.tsx
@@ -8,8 +8,7 @@ import AuthLayout from "@/components/auth/AuthLayout";
 import FormField from "@/components/auth/FormField";
 import ErrorMessage from "@/components/auth/ErrorMessage";
 import AuthButton from "@/components/auth/AuthButton";
-import OAuthButton from "@/components/auth/OAuthButton";
-import { initiateOAuthSignIn } from "@/lib/api/oauth";
+import GoogleButton from "@/components/auth/GoogleButton";
 
 function parseApiError(text: string, fallback: string): string {
   try {
@@ -105,26 +104,6 @@ export default function LoginPage() {
           <p className="text-teal-600">Sign in to your account</p>
         </div>
 
-        <div className="space-y-4">
-          <OAuthButton
-            provider="google"
-            variant="signin"
-            onClick={() => initiateOAuthSignIn("google")}
-            disabled={loading}
-          />
-
-          <div className="relative my-6">
-            <div className="absolute inset-0 flex items-center">
-              <div className="w-full border-t border-gray-300"></div>
-            </div>
-            <div className="relative flex justify-center text-sm">
-              <span className="px-4 bg-white text-gray-500 font-medium">
-                OR
-              </span>
-            </div>
-          </div>
-        </div>
-
         <form onSubmit={handleSubmit} className="space-y-5">
           <ErrorMessage message={error} />
           <FormField
@@ -157,6 +136,24 @@ export default function LoginPage() {
             Sign Up
           </a>
         </p>
+
+        <div className="mt-6">
+          <div className="relative my-6">
+            <div className="absolute inset-0 flex items-center">
+              <div className="w-full border-t border-gray-300"></div>
+            </div>
+            <div className="relative flex justify-center text-sm">
+              <span className="px-4 bg-white text-gray-500 font-medium">
+                OR
+              </span>
+            </div>
+          </div>
+
+          <GoogleButton
+            text="Continue with Google"
+            onClick={() => router.push("/browse")}
+          />
+        </div>
       </div>
     </AuthLayout>
   );
